build: fix npm package files inclusion
We are publishing with yarn 1, which has a problem with negation
patterns in "files" list, causing the npm package to include every
file in the directory.

Since the original reason to expose src files was to address #3652,
which is a client-only concern, we can include the `/client` dir only.
This avoids having to exclude tests in `/node` and also reduces
package size.